refactor(picker): apply picker-crypto review feedback
- registerStock now clears prior error (matches registerCrypto symmetry)
- Add regression test for registry-throws-after-resolve-succeeds path
- Split crypto-select tests into a sibling file to keep both under the
  400-line threshold
